水电配置

+++ src/schema/busiLease/busiLease.contractRenewSchema.js
@@ -220,7 +220,7 @@ module.exports = {
         {
             key: 'roommoney',
             title: '房间租金',
-            dataType: 'int',
+            dataType: 'float',
             placeholder: "请选择合同房间",
             disabled: true
         },
